requirements.txt
flask==2.0.1
pandas==1.3.3
openpyxl==3.0.9
python-docx==0.8.11
requests==2.26.0
werkzeug==2.0.1
